[PR feedback] Implement a helper fixture for temporary workspaces
